Move rmtree and copytree to the internal lib
