+++ lib/python/orchest-internals/_orchest/internals/utils.py
@@ -4,6 +4,7 @@ import json
 import logging
 import os
 import re
+import subprocess
 import time
 import uuid
 from collections import ChainMap
@@ -616,3 +617,54 @@ def docker_has_gpu_capabilities(
             pass
         return False
     return True
+
+
+def rmtree(path, ignore_errors=False):
+    """A wrapped rm -rf.
+
+    If eventlet is being used and it's either patching all modules or
+    patchng subprocess, this function is not going to block the thread.
+
+    Raises:
+        OSError if it failed to copy.
+
+    """
+    exit_code = subprocess.call(f"rm -rf {path}", stderr=subprocess.STDOUT, shell=True)
+    if exit_code != 0 and not ignore_errors:
+        raise OSError(f"Failed to rm {path}: {exit_code}.")
+
+
+def copytree(source: str, target: str):
+    """Copies content from source to target.
+
+    As part of the copying process it ignores patterns from the
+    top-level `.gitignore` in `source`.
+
+    If eventlet is being used and it's either patching all modules or
+    patchng subprocess, this function is not going to block the thread.
+
+    Raises:
+        OSError if it failed to copy.
+
+    """
+    # With a trailing `/` rsync copies the content of the directory
+    # instead of the directory itself.
+    if not source.endswith("/"):
+        source += "/"
+
+    # Construct copy command.
+    # Using rsync with `-W` copies files as a whole which drastically
+    # improves its performance, making it almost as fast as the `cp`
+    # command. The other options (`-aHAX`) are to preserve all kinds
+    # of attributes, e.g. symlinks, `-a` also automatically copies
+    # recursively.
+    copy_cmd = ["rsync", "-aWHAX"]
+    if os.path.isfile(f"{source}.gitignore"):  # source has trailing `/`
+        copy_cmd += [f"--exclude-from={source}.gitignore"]
+    copy_cmd += [f"{source} {target}"]
+
+    exit_code = subprocess.call(
+        " ".join(copy_cmd), stderr=subprocess.STDOUT, shell=True
+    )
+    if exit_code != 0:
+        raise OSError(f"Failed to copy {source} to {target}, :{exit_code}.")
